+++ tests/datasources/test_iri.py
@@ -6,21 +6,18 @@ import cftime
 import numpy as np
 import pytest
 import xarray as xr
-
-# TODO: this will break if there is another conftest
-from conftest import FAKE_AA_DATA_DIR, ISO3
 from xarray.coding.cftimeindex import CFTimeIndex
 
 from aatoolbox import GeoBoundingBox, IriForecastDominant, IriForecastProb
 
-MODULE_BASENAME = "iri"
+DATASOURCE_BASE_DIR = "iri"
 FAKE_IRI_AUTH = "FAKE_IRI_AUTH"
 
 
 @pytest.fixture
 def mock_iri(mock_country_config):
     """Create IRI class with mock country config."""
-    geo_bounding_box = GeoBoundingBox(north=6, south=3.2, east=-2, west=3)
+    geo_bounding_box = GeoBoundingBox(north=6, south=3.2, east=2, west=-3)
 
     def _mock_iri(prob_forecast: bool = True):
         if prob_forecast:
@@ -66,43 +63,47 @@ def mock_download(mocker, mock_iri):
     return _mock_download
 
 
-def test_download_call_prob(mock_download):
+def test_download_call_prob(
+    mock_download, mock_aa_data_dir, mock_country_config
+):
     """Test download for tercile probability forecast."""
     url, filepath = mock_download(prob_forecast=True)
     assert url == (
         "https://iridl.ldeo.columbia.edu/SOURCES/.IRI/.FD/"
         ".NMME_Seasonal_Forecast/"
-        ".Precipitation_ELR/.prob/X/%283.0%29%28-2.0%29RANGEEDGES/"
+        ".Precipitation_ELR/.prob/X/%28-3.0%29%282.0%29RANGEEDGES/"
         "Y/%286.0%29%283.0%29RANGEEDGES/data.nc"
     )
 
     assert filepath == (
-        Path(FAKE_AA_DATA_DIR) / f"private/raw/{ISO3}/{MODULE_BASENAME}/"
+        mock_aa_data_dir
+        / f"private/raw/{mock_country_config.iso3}/{DATASOURCE_BASE_DIR}/"
         f"abc_iri_forecast_seasonal_precipitation_"
-        f"tercile_prob_Np6Sp3Em2Wp3.nc"
+        f"tercile_prob_Np6Sp3Ep2Wm3.nc"
     )
 
 
-def test_download_call_dominant(mock_download):
+def test_download_call_dominant(
+    mock_download, mock_aa_data_dir, mock_country_config
+):
     """Test download for dominant tercile forecast."""
     url, filepath = mock_download(prob_forecast=False)
     assert url == (
         "https://iridl.ldeo.columbia.edu/SOURCES/.IRI/.FD/"
         ".NMME_Seasonal_Forecast/"
-        ".Precipitation_ELR/.dominant/X/%283.0%29%28-2.0%29RANGEEDGES/"
+        ".Precipitation_ELR/.dominant/X/%28-3.0%29%282.0%29RANGEEDGES/"
         "Y/%286.0%29%283.0%29RANGEEDGES/data.nc"
     )
 
     assert filepath == (
-        Path(FAKE_AA_DATA_DIR) / f"private/raw/{ISO3}/{MODULE_BASENAME}/"
+        mock_aa_data_dir
+        / f"private/raw/{mock_country_config.iso3}/{DATASOURCE_BASE_DIR}/"
         f"abc_iri_forecast_seasonal_"
-        f"precipitation_tercile_dominant_Np6Sp3Em2Wp3.nc"
+        f"precipitation_tercile_dominant_Np6Sp3Ep2Wm3.nc"
     )
 
 
-# TODO: need to use a tmp dir but will copy
-# that from glofas once that is ready :)
-def test_process(mocker, mock_iri):
+def test_process(mocker, mock_iri, mock_aa_data_dir, mock_country_config):
     """Test process for IRI forecast."""
     ds = xr.DataArray(
         np.reshape(a=np.arange(16), newshape=(2, 2, 2, 2)),
@@ -130,9 +131,10 @@ def test_process(mocker, mock_iri):
 
     processed_path = iri.process()
     assert processed_path == (
-        Path(FAKE_AA_DATA_DIR) / f"private/processed/{ISO3}/{MODULE_BASENAME}/"
-        f"{ISO3}_iri_forecast_seasonal_precipitation_"
-        f"tercile_prob_Np6Sp3Em2Wp3.nc"
+        mock_aa_data_dir / f"private/processed/{mock_country_config.iso3}/"
+        f"{DATASOURCE_BASE_DIR}/{mock_country_config.iso3}_"
+        f"iri_forecast_seasonal_precipitation_"
+        f"tercile_prob_Np6Sp3Ep2Wm3.nc"
     )
 
     da_processed = xr.load_dataset(processed_path)
@@ -175,7 +177,13 @@ def mock_xr_load_dataset(mocker):
     )
 
 
-def test_iri_load(mocker, mock_xr_load_dataset, mock_iri):
+def test_iri_load(
+    mocker,
+    mock_xr_load_dataset,
+    mock_iri,
+    mock_aa_data_dir,
+    mock_country_config,
+):
     """Test that load_codab calls the HDX API to download."""
     mocker.patch("aatoolbox.datasources.iri.iri_seasonal_forecast._download")
 
@@ -184,10 +192,11 @@ def test_iri_load(mocker, mock_xr_load_dataset, mock_iri):
     mock_xr_load_dataset.assert_has_calls(
         [
             mocker.call(
-                Path(FAKE_AA_DATA_DIR)
-                / f"private/processed/{ISO3}/{MODULE_BASENAME}/"
-                f"{ISO3}_iri_forecast_seasonal_precipitation_"
-                f"tercile_prob_Np6Sp3Em2Wp3.nc"
+                mock_aa_data_dir
+                / f"private/processed/{mock_country_config.iso3}/"
+                f"{DATASOURCE_BASE_DIR}/{mock_country_config.iso3}"
+                f"_iri_forecast_seasonal_precipitation_"
+                f"tercile_prob_Np6Sp3Ep2Wm3.nc"
             ),
         ]
     )