+++ CHANGELOG.md
@@ -17,6 +17,12 @@ and this project adheres to
 - Implemented downloading and processing for IRI seasonal precipitation forecast
 - Added config for DRC (of which the iso3 is COD)
 
+### Changed
+
+- `DataSource` is now an abstract base class with required `download`, `process`
+  and `load` methods
+- `GeoBoundingBox.round` returns `GeoBoundingBox` instance (instead of being in place)
+
 ### Removed
 
 - `Pipeline` class no longer used as main API
@@ -28,6 +34,7 @@ and this project adheres to
   (using tags instead)
 - HDX API now uses "prod" server, and version >= 5.5.8 to avoid download error
 - COD AB dataset URLs on HDX are standardized
+- `GeoBoundingBox` won't allow north < south or east < west
 
 ## [0.3.1] - 2022-01-06